working on changing active business and active business id. Also built file structure for financial data
