+++ src/components/financial/businessFinancials.js
@@ -1,14 +1,23 @@
 import React, {Component} from 'react';
 
+import BalanceSheet from './BalanceSheet.js';
+import IncomeStatement from './IncomeStatement.js';
+
+
 export default class FinancialMarketingSalesDataInput extends Component {
   constructor(props) {
     super(props);
   }
 
+  componentDidMount(){
+    
+  }
+
+
   render(){
     return(
       <div>
-        <p></p>
+        <IncomeStatement/>
       </div>
     )
   }
