+++ src/components/memberHome/activeBusiness.js
@@ -3,12 +3,17 @@ import React, {Component} from 'react';
 export default class ActiveBusiness extends Component {
   constructor(props){
     super(props);
+    this.state={
+      active: '',
+      activeBid: ''
+    }
   }
 
   render(){
+
     return(
       <div className="userBusiness-name">
-        <h1 className="active-business">{this.props.activBusiness}</h1>
+        <button className="btn btn-primary" onClick={this.handleActiveClick.bind(this)}>Change Active Business</button>
       </div>
     )
   }
