+++ src/styles/memberDesktop.css
@@ -11,7 +11,7 @@
 }
 
 .header-container {
-  flex-basis: 30%;
+  flex-basis: 40%;
   background-image: url('./images/3.png');
   height: 125px;
   background-size: contain;
@@ -40,4 +40,43 @@
   font-size: 1.5em;
   border-radius: 0;
   background-color: #5282BA;
+}
+
+
+/*
+MemberHomeBox styling
+*/
+
+.memberHome-container {
+  display: flex;
+  flex-direction: row;
+  flex-wrap:
+}
+
+.title-content-container {
+  flex-basis: 85%;
+  height: 80px;
+  text-align: center;
+}
+
+.settings-container {
+  flex-basis: 15%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.biz-header {
+  height: 80px;
+  text-align: center;
+}
+
+/*
+Add Biz styling
+*/
+.form-control {
+  width: 100%;
+}
+.add-biz-container {
+  width: 100%;
 }
